+++ stateproofs/stateproofverification/stateproof/verifier.go
@@ -1,4 +1,4 @@
-package stateprooftypes
+package stateproof
 
 import (
 	"errors"
