+++ stateproofs/stateproofverification/stateproof/weights.go
@@ -1,4 +1,4 @@
-package stateprooftypes
+package stateproof
 
 import (
 	"errors"
